@@ -5,30 +5,30 @@ import 'media_query_observer.dart';
 
 class NestedApp extends StatefulWidget {
   const NestedApp({
-    Key key,
-    @required this.route,
-    @required this.size,
-    @required this.customWidget,
-    @required this.materialApp,
-    @required this.cupertinoApp,
-    @required this.widgetsApp,
-    @required this.child,
+    Key? key,
+    required this.route,
+    required this.size,
+    required this.customWidget,
+    required this.materialApp,
+    required this.cupertinoApp,
+    required this.widgetsApp,
+    required this.child,
   }) : super(key: key);
 
-  final Widget child;
-  final CupertinoApp cupertinoApp;
+  final Widget? child;
   final bool customWidget;
-  final MaterialApp materialApp;
-  final RouteSettings route;
-  final Size size;
-  final WidgetsApp widgetsApp;
+  final CupertinoApp? cupertinoApp;
+  final MaterialApp? materialApp;
+  final WidgetsApp? widgetsApp;
+  final RouteSettings? route;
+  final Size? size;
 
   @override
   _NestedAppState createState() => _NestedAppState();
 }
 
 class _NestedAppState extends State<NestedApp> {
-  GlobalObjectKey<NavigatorState> _navKey;
+  late GlobalObjectKey<NavigatorState> _navKey;
 
   @override
   void initState() {
@@ -38,11 +38,11 @@ class _NestedAppState extends State<NestedApp> {
 
   void setup() {
     _navKey = GlobalObjectKey<NavigatorState>(this);
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (widget?.route != null) {
-        _navKey.currentState.pushReplacementNamed(
-          widget.route.name,
-          arguments: widget.route?.arguments,
+    WidgetsBinding.instance?.addPostFrameCallback((_) {
+      if (widget.route != null) {
+        _navKey.currentState?.pushReplacementNamed(
+          widget.route!.name!,
+          arguments: widget.route!.arguments,
         );
       }
     });
@@ -50,14 +50,14 @@ class _NestedAppState extends State<NestedApp> {
 
   @override
   void didUpdateWidget(NestedApp oldWidget) {
-    if (oldWidget?.materialApp != widget?.materialApp) {
-      if (mounted) setState(() => setup());
+    if (oldWidget.materialApp != widget.materialApp) {
+      if (mounted) setState(setup);
     }
-    if (oldWidget?.cupertinoApp != widget?.cupertinoApp) {
-      if (mounted) setState(() => setup());
+    if (oldWidget.cupertinoApp != widget.cupertinoApp) {
+      if (mounted) setState(setup);
     }
-    if (oldWidget?.widgetsApp != widget?.widgetsApp) {
-      if (mounted) setState(() => setup());
+    if (oldWidget.widgetsApp != widget.widgetsApp) {
+      if (mounted) setState(setup);
     }
     // if (oldWidget?.route != widget?.route) {
     //   if (mounted) setState(() => setup());
@@ -66,65 +66,75 @@ class _NestedAppState extends State<NestedApp> {
   }
 
   Map<String, WidgetBuilder> get routes {
-    if (widget.materialApp != null) return widget.materialApp?.routes;
-    if (widget.widgetsApp != null) return widget.widgetsApp?.routes;
-    if (widget.cupertinoApp != null) return widget.cupertinoApp?.routes;
-    return null;
+    if (widget.materialApp != null) return widget.materialApp!.routes!;
+    if (widget.widgetsApp != null) return widget.widgetsApp!.routes!;
+    if (widget.cupertinoApp != null) return widget.cupertinoApp!.routes!;
+    return const <String, WidgetBuilder>{};
   }
 
-  Locale get locale {
+  Locale? get locale {
     if (widget.materialApp != null) return widget.materialApp?.locale;
     if (widget.widgetsApp != null) return widget.widgetsApp?.locale;
     if (widget.cupertinoApp != null) return widget.cupertinoApp?.locale;
     return null;
   }
 
-  Iterable<LocalizationsDelegate<dynamic>> get localizationsDelegates {
-    if (widget.materialApp != null)
+  Iterable<LocalizationsDelegate<dynamic>>? get localizationsDelegates {
+    if (widget.materialApp != null) {
       return widget.materialApp?.localizationsDelegates;
-    if (widget.widgetsApp != null)
+    }
+    if (widget.widgetsApp != null) {
       return widget.widgetsApp?.localizationsDelegates;
-    if (widget.cupertinoApp != null)
+    }
+    if (widget.cupertinoApp != null) {
       return widget.cupertinoApp?.localizationsDelegates;
+    }
     return null;
   }
 
   Iterable<Locale> get supportedLocales {
-    if (widget.materialApp != null) return widget.materialApp?.supportedLocales;
-    if (widget.widgetsApp != null) return widget.widgetsApp?.supportedLocales;
-    if (widget.cupertinoApp != null)
-      return widget.cupertinoApp?.supportedLocales;
-    return null;
+    if (widget.materialApp != null) return widget.materialApp!.supportedLocales;
+    if (widget.widgetsApp != null) return widget.widgetsApp!.supportedLocales;
+    if (widget.cupertinoApp != null) {
+      return widget.cupertinoApp!.supportedLocales;
+    }
+    return const <Locale>[Locale('en', 'US')];
   }
 
-  LocaleResolutionCallback get localeResolutionCallback {
-    if (widget.materialApp != null)
+  LocaleResolutionCallback? get localeResolutionCallback {
+    if (widget.materialApp != null) {
       return widget.materialApp?.localeResolutionCallback;
-    if (widget.widgetsApp != null)
+    }
+    if (widget.widgetsApp != null) {
       return widget.widgetsApp?.localeResolutionCallback;
-    if (widget.cupertinoApp != null)
+    }
+    if (widget.cupertinoApp != null) {
       return widget.cupertinoApp?.localeResolutionCallback;
+    }
     return null;
   }
 
-  LocaleListResolutionCallback get localeListResolutionCallback {
-    if (widget.materialApp != null)
+  LocaleListResolutionCallback? get localeListResolutionCallback {
+    if (widget.materialApp != null) {
       return widget.materialApp?.localeListResolutionCallback;
-    if (widget.widgetsApp != null)
+    }
+    if (widget.widgetsApp != null) {
       return widget.widgetsApp?.localeListResolutionCallback;
-    if (widget.cupertinoApp != null)
+    }
+    if (widget.cupertinoApp != null) {
       return widget.cupertinoApp?.localeListResolutionCallback;
+    }
     return null;
   }
 
-  Map<Type, Action<Intent>> get actions {
+  Map<Type, Action<Intent>>? get actions {
     if (widget.materialApp != null) return widget.materialApp?.actions;
     if (widget.widgetsApp != null) return widget.widgetsApp?.actions;
     if (widget.cupertinoApp != null) return widget.cupertinoApp?.actions;
     return null;
   }
 
-  Map<LogicalKeySet, Intent> get shortcuts {
+  Map<LogicalKeySet, Intent>? get shortcuts {
     if (widget.materialApp != null) return widget.materialApp?.shortcuts;
     if (widget.widgetsApp != null) return widget.widgetsApp?.shortcuts;
     if (widget.cupertinoApp != null) return widget.cupertinoApp?.shortcuts;
@@ -132,33 +142,36 @@ class _NestedAppState extends State<NestedApp> {
   }
 
   String get title {
-    if (widget.materialApp != null) return widget.materialApp?.title;
-    if (widget.widgetsApp != null) return widget.widgetsApp?.title;
-    if (widget.cupertinoApp != null) return widget.cupertinoApp?.title;
-    return null;
+    if (widget.materialApp != null) return widget.materialApp!.title;
+    if (widget.widgetsApp != null) return widget.widgetsApp!.title;
+    if (widget.cupertinoApp != null) return widget.cupertinoApp!.title;
+    return '';
   }
 
-  String Function(BuildContext) get onGenerateTitle {
+  String Function(BuildContext)? get onGenerateTitle {
     if (widget.materialApp != null) return widget.materialApp?.onGenerateTitle;
     if (widget.widgetsApp != null) return widget.widgetsApp?.onGenerateTitle;
-    if (widget.cupertinoApp != null)
+    if (widget.cupertinoApp != null) {
       return widget.cupertinoApp?.onGenerateTitle;
+    }
     return null;
   }
 
   Widget builder(BuildContext context, Widget child) {
-    if (widget.materialApp?.builder != null)
-      return widget.materialApp?.builder(context, child);
-    if (widget.widgetsApp?.builder != null)
-      return widget.widgetsApp?.builder(context, child);
-    if (widget.cupertinoApp?.builder != null)
-      return widget.cupertinoApp?.builder(context, child);
+    if (widget.materialApp?.builder != null) {
+      return widget.materialApp!.builder!(context, child);
+    }
+    if (widget.widgetsApp?.builder != null) {
+      return widget.widgetsApp!.builder!(context, child);
+    }
+    if (widget.cupertinoApp?.builder != null) {
+      return widget.cupertinoApp!.builder!(context, child);
+    }
     return child;
   }
 
   @override
   Widget build(BuildContext context) {
-    assert(widget.child != null || widget.route != null);
     if (widget.customWidget) {
       return MaterialApp(
         theme: widget.materialApp?.theme ?? ThemeData.light(),
@@ -168,7 +181,7 @@ class _NestedAppState extends State<NestedApp> {
         home: widget.child,
         builder: (context, val) => MediaQueryObserver(
           data: MediaQuery.of(context).copyWith(size: widget.size),
-          child: builder(context, val),
+          child: builder(context, val ?? Container()),
         ),
         color: widget.materialApp?.color,
         locale: locale,
@@ -197,7 +210,7 @@ class _NestedAppState extends State<NestedApp> {
         home: widget.child,
         builder: (context, val) => MediaQueryObserver(
           data: MediaQuery.of(context).copyWith(size: widget.size),
-          child: builder(context, val),
+          child: builder(context, val ?? Container()),
         ),
         // initialRoute: route?.name,
         color: widget.materialApp?.color,
@@ -215,7 +228,7 @@ class _NestedAppState extends State<NestedApp> {
     if (widget.cupertinoApp != null) {
       return CupertinoApp(
         navigatorKey: _navKey,
-        theme: widget.cupertinoApp?.theme ?? CupertinoThemeData(),
+        theme: widget.cupertinoApp?.theme ?? const CupertinoThemeData(),
         debugShowCheckedModeBanner: false,
         routes: routes,
         onGenerateRoute: widget.cupertinoApp?.onGenerateRoute,
@@ -224,7 +237,7 @@ class _NestedAppState extends State<NestedApp> {
         home: widget.child,
         builder: (context, val) => MediaQueryObserver(
           data: MediaQuery.of(context).copyWith(size: widget.size),
-          child: builder(context, val),
+          child: builder(context, val ?? Container()),
         ),
         // initialRoute: route?.name,
         color: widget.cupertinoApp?.color,
@@ -250,10 +263,10 @@ class _NestedAppState extends State<NestedApp> {
         home: widget.child,
         builder: (context, val) => MediaQueryObserver(
           data: MediaQuery.of(context).copyWith(size: widget.size),
-          child: builder(context, val),
+          child: builder(context, val ?? Container()),
         ),
         // initialRoute: route?.name,
-        color: widget.widgetsApp?.color,
+        color: widget.widgetsApp?.color ?? Colors.transparent,
         locale: locale,
         localizationsDelegates: localizationsDelegates,
         supportedLocales: supportedLocales,
@@ -265,6 +278,6 @@ class _NestedAppState extends State<NestedApp> {
         onGenerateTitle: onGenerateTitle,
       );
     }
-    return widget.child;
+    return widget.child ?? Container();
   }
 }
