@@ -2,16 +2,14 @@
 // Generated file. Do not edit.
 //
 
-// ignore: unused_import
-import 'dart:ui';
+// ignore_for_file: lines_longer_than_80_chars
 
 import 'package:shared_preferences_web/shared_preferences_web.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 // ignore: public_member_api_docs
-void registerPlugins(PluginRegistry registry) {
-  SharedPreferencesPlugin.registerWith(
-      registry.registrarFor(SharedPreferencesPlugin));
-  registry.registerMessageHandler();
+void registerPlugins(Registrar registrar) {
+  SharedPreferencesPlugin.registerWith(registrar);
+  registrar.registerMessageHandler();
 }
